Replace run.py with the finished newrun.py
